Split out prefill
